+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-## 0.0.1
-- Initial version, created gbk map from ICU.
+## 0.1.0
+Initial version, implement GBK codec.
+- support for GBK to String encode and decode.
+- support stream use case in File reading/writing & HttpClient requests&response.